Fix a bug in client filter
The bug was that the select client selection box is pre-selecting a
client even if the user didn't put select anything from it at all. It
was due to the localStorage which I have setup before. Instead of
relying on the localStorage to determine if I have to populate the
client filter selectbox, I should be using the currentQuery from
drupalSettings instead.