Refactor the logic
This refactoring made was to the logic a bit, in terms of how facebook
pages are being fetched.

This time around, the fb pages are now being stored as an
ic_fb_page_entity type. Whenever a user lands on the dashboard page, fb
pages are collected. If the fb page already exists, the page access
token of that is gets updated. If the fb page does not exists yet, then
the entity is created, storing the page id, name and page access token.

getFbPageInsights() and getTopFiveContents() now relies on the client
query from the url as the value of it will determine which page id and
access token to use (The site admin should assign the corresponding
fb page to the user/client it's creating. This is how it's going to work)