Refactor IcFbService.php
Refactor the logic regarding on FB Conversations. First, the
conversation messages are being stored as a facebook_message entity
(when the client url parameter is available).
I also added a checking to see if the entity is already created so that
we don't create duplicates.

Then I also refactor the way on how to count the fb message keywords.
Since the fb conversation messages are now entities, we are now
queriying them from the database instead of fetching them from the
conversation fb api.